backend_app/requirements.txt
aiohttp==3.6.2
amqp==2.5.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.1.0
boto3==1.16.23
botocore==1.19.23
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cpe==1.2.1
cvsslib==0.6.0
deepdiff==4.2.0
Django==3.0.11
django-annoying==0.10.6
django-celery-beat==2.0.0
django-celery-results==2.0.0
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-extensions==2.2.8
django-filter==2.2.0
django-organizations==1.1.2
django-simple-history==2.12.0
django-timezone-field==4.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
future==0.18.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.4.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
kombu==4.6.7
MarkupSafe==1.1.1
mongoengine==0.19.1
more-itertools==8.1.0
multidict==4.7.5
oauthlib==3.1.0
openapi-codec==1.3.2
ordered-set==3.1.1
packaging==20.3
psycopg2==2.8.4
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.6
python-crontab==2.4.0
python-dateutil==2.8.1
python-twitter==3.5
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.13.0
slackclient==2.5.0
sqlparse==0.3.0
supervisor==4.1.0
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
yarl==1.4.2
zipp==0.6.0
tqdm
